+++ lib/adminlte2-rails.rb
@@ -1,2 +1,4 @@
 require 'adminlte2/rails'
 require 'bootstrap-sass'
+require 'font-awesome-rails'
+require 'base_generator'